Prep for Part 2 of assignment
To address the alignment issues brought up in class, the GunRegistrationPoint object is now zeroed out.

To test alignment with your weapon, just make it a child of this object (A in the screenshot) and reset the position, rotation, and scale of your weapon. That is how the system will place it.

The Player object is no longer a prefab. This will allow you to keep your weapon as a child of the GunRegistrationPoint and apply the weapon’s prefab without getting mixed up with the Player prefab.

The RigidbodyFirstPersonController has been modified to allow it to be pushed around by physics like in the Rocket Jumping weapon we saw in class.

There are now simple visuals in the Player so we will have a point of reference.

You can now access your Player’s camera via a new field on Weapon called playerCamera. From inside your own Weapon subclass script, you can just type .playerCamera and you can count on it being set as long as your weapon is active.